refactor: bilibili-share-to-friends 下沉发送逻辑
