+++ scripts/bilibili-share-to-friends/src/components/dialog-footer/DialogFooter.jsx
@@ -1,13 +1,49 @@
+import { useEffect, useState } from "preact/hooks";
+
+import { assertLogin, sendVideoText } from "../../api.js";
 import { SCRIPT_ID } from "../../constants.js";
 import "./style.css";
 
 export const DialogFooter = ({
-  sending = false,
-  canSend = false,
+  video,
+  selectedUser,
   showCloseOnly = false,
   onClose,
-  onSend = () => {},
+  onSendingChange = () => {},
+  onSendSuccess = () => {},
+  onSendError = () => {},
 }) => {
+  const [sending, setSending] = useState(false);
+
+  useEffect(() => {
+    onSendingChange(sending);
+  }, [onSendingChange, sending]);
+
+  const handleSend = async () => {
+    if (!selectedUser || sending) {
+      return;
+    }
+    setSending(true);
+    onSendError("");
+    try {
+      const login = await assertLogin();
+      await sendVideoText({
+        nav: login.nav,
+        csrf: login.csrf,
+        video,
+        receiver: selectedUser,
+      });
+      onSendSuccess({
+        message: `已将视频链接发送给 ${selectedUser.name}。`,
+        isError: false,
+      });
+    } catch (sendError) {
+      onSendError(sendError.message);
+    } finally {
+      setSending(false);
+    }
+  };
+
   if (showCloseOnly) {
     return (
       <div className={`${SCRIPT_ID}-footer`}>
@@ -30,8 +66,8 @@ export const DialogFooter = ({
       <button
         className={`${SCRIPT_ID}-btn ${SCRIPT_ID}-btn-primary`}
         type="button"
-        disabled={sending || !canSend}
-        onClick={onSend}
+        disabled={sending || !selectedUser}
+        onClick={handleSend}
       >
         {sending ? "发送中" : "发送"}
       </button>
